requirements.txt
defusedxml==0.4.1
Django==1.9.1
django-allauth==0.24.1
django-braces==1.8.1
django-crispy-forms==1.6.0
django-pagination==1.0.7
oauthlib==1.0.3
Pillow==3.1.0
psycopg2==2.6.1
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.24.0
